Update on Medication and Lab Result API Endpoint

+++ api/v1/views/medications.py
@@ -2,7 +2,6 @@
 """
 This module creates view for Medication objects
 """
-
 from flask import jsonify, request, abort
 from api.v1.views import app_views
 from models import storage
@@ -11,33 +10,18 @@ from models.patient import Patient
 from models.medication import Medication
 
 
-@app_views.route('/patient/<patient_id>/medications', methods=['GET'], strict_slashes=False)
-def get_patient_medications(patient_id):
+@app_views.route('/patient/<patient_id>/medication', methods=['GET'], strict_slashes=False)
+def get_patient_medication(patient_id):
     """Retrieves a Patient Medication based on the patient_id"""
     patient = storage.get(Patient, patient_id)
     if not patient:
-        # abort(400, "Patient does not exist")
-        return jsonify({"Response": "Patient does not exist"}), 400
+        abort(400, "Patient does not exist")
 
-    # medication = storage._DBStorage__session.query(Medication).filter_by(patient_id=patient_id).first()
-    all_meds = storage.all(Medication).values()
-    medication = [meds.to_dict() for meds in all_meds if meds.patient_id == patient_id]
-
+    medication = storage._DBStorage__session.query(Medication).filter_by(patient_id=patient_id).first()
     if not medication:
-        # abort(400, "No Medication for the specified patient!")
-        return jsonify({"Response": "No Medication for the specified patient"}), 400
+        abort(400, "No medication found for the specified Patient")
 
-    return (jsonify(medication)), 201
-
-
-@app_views.route('/patient/medication/<medication_id>', methods=['GET'], strict_slashes=False)
-def get_a_patient_medication(medication_id):
-    """Retrieves a Patient Medication based on the patient_id"""
-    medication = storage.get(Medication, medication_id)
-    if not medication:
-        return jsonify({"Response": "Medication does not exist"}), 400
-
-    return (jsonify(medication.to_dict())), 201
+    return jsonify(medication.to_dict())
 
 
 @app_views.route('/doctor/<doctor_id>/patient/<patient_id>/medication', methods=['POST'], strict_slashes=False)
@@ -47,9 +31,9 @@ def create_patient_medication(doctor_id, patient_id):
     patient = storage.get(Patient, patient_id)
 
     if not doctor:
-        return jsonify({"Response": "Doctor does not exist!"})
+        abort(400, "Doctor does not exist!")
     if not patient:
-        return jsonify({"Response": "Patient does not exist!"})
+        abort(400, "Patient does not exist!")
     
     data = request.get_json()
     if not data:
@@ -58,7 +42,12 @@ def create_patient_medication(doctor_id, patient_id):
     required_fields = ['medicine_name', 'dosage', 'frequency', 'duration']
     for field in required_fields:
         if field not in data:
-            return jsonify({"Response": f"Missing {field}"})
+            abort(400, f"Missing {field}")
+
+    existing_record = storage._DBStorage__session.query(Medication).filter_by(patient_id=patient_id).first()
+    if existing_record:
+        abort(400, "Patient already has a medication")
+
     try:
         data['doctor_id'] = doctor_id
         data['patient_id'] = patient_id
@@ -67,15 +56,24 @@ def create_patient_medication(doctor_id, patient_id):
         storage.save()
     except Exception as e:
         abort(500, f"An error occured while saving the Patient: {str(e)}")
+
     return (jsonify(medication.to_dict())), 201
 
 
-@app_views.route('/patient/medication/<medication_id>', methods=['PUT'], strict_slashes=False)
-def update_patient_medication(medication_id):
-    """Updates a Patient Medication based on the medication_id"""
-    medication = storage.get(Medication, medication_id)
+@app_views.route('/doctor/<doctor_id>/patient/<patient_id>/medication', methods=['PUT'], strict_slashes=False)
+def update_patient_medication(doctor_id, patient_id):
+    """Updates a Patient Medication based on patient_id"""
+    doctor = storage.get(Doctor, doctor_id)
+    if not doctor:
+        abort(400, "Doctor does not exist")
+
+    patient = storage.get(Patient, patient_id)
+    if not patient:
+        abort(400, "Patient does not exist")
+
+    medication = storage._DBStorage__session.query(Medication).filter_by(patient_id=patient_id).first()
     if not medication:
-        abort(404, "Medication not found")
+        abort(400, " Medication record not found for the specified Patient")
 
     data = request.get_json()
     if not data:
@@ -95,10 +93,21 @@ def update_patient_medication(medication_id):
     return jsonify(medication.to_dict()), 201
 
 
-@app_views.route('/patient/medication/<medication_id>', methods=['DELETE'], strict_slashes=False)
-def delete_patient_medication(medication_id):
+@app_views.route('/doctor/<doctor_id>/patient/<patient_id>/medication', methods=['DELETE'], strict_slashes=False)
+def delete_patient_medication(doctor_id, patient_id):
     """Deletes a Patient Mediaction based on the patient_id"""
-    medication = storage.get(Medication, medication_id)
+    doctor = storage.get(Doctor, doctor_id)
+    if not doctor:
+        abort(400, "Doctor does not exist")
+
+    patient = storage.get(Patient, patient_id)
+    if not patient:
+        abort(400, "Patient does not exist")
+
+    medication = storage._DBStorage__session.query(Medication).filter_by(patient_id=patient_id).first()
+    if not medication:
+        abort(400, "Medication not found for the specified Patient")
+
     storage.delete(medication)
     storage.save()
     return jsonify({}), 200
